separated conversation modal from alumni networking
